fix "Foreign key constraint is incorrectly formed" problem in migrations

+++ database/migrations/2020_09_22_094213_create_brands_table.php
@@ -16,7 +16,7 @@ class CreateBrandsTable extends Migration
         Schema::create('brands', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('photo_id');
+            $table->unsignedBigInteger('photo_id');
             $table->timestamps();
 
             $table->foreign('photo_id')
